Factors gene_annotation and reg_effects into two files each
Pulls the "Features" models out of gene_annotation and the "DNARegion"
model out of reg_effects. They were where they were for historical
reasons, but they now should be separate in more obvious locations.
