+++ cegs_portal/search/models/__init__.py
@@ -1,8 +1,10 @@
+from .dna_region import DNARegion
 from .experiment import Experiment, ExperimentDataFile
 from .facets import Facet, FacetedModel, FacetValue
+from .features import Feature, FeatureAssembly
 from .file import File
-from .gene_annotation import Feature, FeatureAssembly, GencodeAnnotation, GencodeRegion
-from .reg_effects import DNARegion, EffectDirectionType, RegulatoryEffect
+from .gene_annotation import GencodeAnnotation, GencodeRegion
+from .reg_effects import EffectDirectionType, RegulatoryEffect
 from .searchable import Searchable
 from .utils import ChromosomeLocation, QueryToken
 from .variants import Subject, Variant, VCFEntry, VCFFile
